Move from 32 bit floats to 64 bit. Increases runtime to 51ms (FFTW) and 79ms (RustFFT)

+++ caf_rust/benches/caf_bench.rs
@@ -16,7 +16,7 @@ mod caf_benches {
         // -100Hz to 100Hz, 0.5Hz step
         let mut shifts = Vec::new();
         for shift_millihz in (-100000..100000).step_by(500) {
-            let shift = (shift_millihz as f32) / 1e3;
+            let shift = (shift_millihz as f64) / 1e3;
             shifts.push(shift);
         }
